chore: Move into a docker
Move server code to a docker and make up Dockerfile for setting up and running.
The docker uses pytorch of 2.6.0 and cuda 11.8 for the future tasks with GPU, and installs requisite packages of fastAPI and uvicorn for opening server and providing APIs. It might be needed to specify the version of the requisite packages.

README file is also updated according to these changes.