impl: Enhance HTML works
Using FakeDB, the service now provides image storing and searching.
The main page shows the number of images stored in DB, input elements for uploading a image and querying a text.
When you upload a image or query a text, instead of printing a feature vector of it, the service works like this:
- When you upload a image, it responds just with success/failure text.
- When you query a text, it responds with a few most-matching images. (5 images or fewer if the DB contains that much)