[chore] Make a workflow / Change required files to build & run / Prevent bots
Make a workflow (for the first time) to automatically build & push the image, connect to EC2, pull the image and restart the server.
For this, some variables of repository secrets are added in repository settings.
Also, the image would not contain the whole module file. It should exist inside the folder named "model" in working directory, and be accessed by mounting the folder into the container at creating, like ".env.local".
Accordingly, the model name string in src/core.py is replaced to "model" folder path.

Add "robots.txt" to prevents bots from requesting our server.